Fix bug with async instantiation
There was a bug where you could create multiple instances of a
component.

Before, we only registered the instance after we created it, meaning
that any other call to '_get' in between the initial call and resolution
would result in another instance being created.

The solution is to synchronously register the promise for the instance,
so any subsequent requests will use that.